+++ .gitignore
@@ -30,4 +30,6 @@ Carthage
 # Note: if you ignore the Pods directory, make sure to uncomment
 # `pod install` in .travis.yml
 #
-# Pods/
+Pods/
+Podfile.lock
+*.xcworkspacedata
